Resolve segundo exercicio
